move session token to separate class and add session data parsing
